requirements.txt
# pip freeze > requirements.txt to save packets names
# pip install -r requirements.txt to load all packets

alembic==1.4.3
Babel==2.9.0
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.0.0
dominate==2.6.0
elasticsearch==7.10.1
email-validator==1.1.2
flake8==3.8.4
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
guess-language-spirit==0.5.3
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
langdetect==1.0.8
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.22
nltk==3.5
pkg-resources==0.0.0
protobuf==3.14.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
regex==2020.11.13
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.20
textblob==0.15.3
tqdm==4.55.1
urllib3==1.26.2
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
